refactor: move version to front of route
