+++ backend/api/routers/v1/index.ts
@@ -2,15 +2,15 @@ import express from 'express';
 
 import { ControllerMap, MiddlewareMap } from '@/lib';
 
-import V1Router from './v1';
-import V2BetaRouter from './v2alpha';
+import StatefulRouter from './stateful';
+import StatelessRouter from './stateless';
 
 export default (middlewares: MiddlewareMap, controllers: ControllerMap) => {
   const router = express.Router();
 
   router.get('/health', (_, res) => res.send(`${process.env.NODE_ENV} Healthy`));
-  router.use('/v1', V1Router(middlewares, controllers));
-  router.use('/v2beta', V2BetaRouter(middlewares, controllers));
+  router.use('/stateful', StatefulRouter(middlewares, controllers));
+  router.use('/stateless', StatelessRouter(middlewares, controllers));
 
   return router;
 };
